==> analyses/0.1-preliminary_stats.ipynb <==
# Preliminary stats on metadata for ContNeXt

import getpass
import json
import os
import sys
import time
from urllib import request

import pandas as pd
from tqdm import tqdm

getpass.getuser()

sys.version

time.asctime()

# replace here the location of the external data dir
data_dir = os.path.join(os.path.expanduser("~"), "contnext_data", "data")

### Load Gemma dump

df = pd.read_table(os.path.join(data_dir, "metadata", "gemma_dump_ontologies.tsv"))
df = df.loc[df["class.Name"].isin(['organism part', 'cell type', 'cell line'])]
df

### add platform column

def get_platform(dataset):
    if dataset == "McLean Hippocampus": ### workaround for this ds
        dataset = "670"                 ### workaround
    if " " in dataset:
        dataset = dataset.replace(" ", "%20") ##this is supposed to be the workaround for spaces, but it doesn't work with the above
    url = f"https://gemma.msl.ubc.ca/rest/v2/datasets/{dataset}/platforms"
    webURL = request.urlopen(url)
    data = webURL.read()
    encoding = webURL.info().get_content_charset('utf-8')
    contents = json.loads(data.decode(encoding))
    platforms = []
    for platform in contents["data"]:
        platforms.append(platform["shortName"])
    return ", ".join(platforms)

try:
    with open(os.path.join(data_dir, "metadata", 'platforms.json'), 'r') as f:
        platform_dict = json.load(f)
except:
    platform_dict = {}
    
dict_changed = False

datasets = df["experiment.ShortName"]
platforms = []
error_datasets = []

for dataset in tqdm(datasets):
    if dataset in platform_dict:
        platforms.append(platform_dict[dataset])
        continue
    try:
        pltfm = get_platform(dataset)
        if not pltfm:
            pltfm = pd.NA
    except:
        pltfm = pd.NA
        error_datasets.append(dataset)
    platforms.append(pltfm)
    if not pd.isnull(pltfm):
        dict_changed = True
        platform_dict[dataset] = pltfm
    
if dict_changed:
    with open(os.path.join(data_dir, "metadata", 'platforms.json'), 'w') as f:
        json.dump(platform_dict, f)

len(platforms), len(datasets)

df['platform.ShortName'] = platforms
df_subset = df.dropna()

df_subset.head(20)

df_human_only = df_subset.loc[df_subset["taxon.Name"] == "human"]
df_GPL570 = df_subset[df_subset['platform.ShortName'].str.contains("GPL570")]

### get sample metadata

path = os.path.join(data_dir, "metadata", "dataset_sample_metadata")

if not os.path.exists(path):
    os.makedirs(path)
    
def get_sample_metadata(dataset, path):
    if dataset == "McLean Hippocampus": ### workaround
        dataset = "670"                 ### workaround
    if os.path.exists(os.path.join(path,dataset,'sample_metadata.json')):
        with open(os.path.join(path,dataset,'sample_metadata.json'), 'r') as f:
            return json.load(f)
    url = f"https://gemma.msl.ubc.ca/rest/v2/datasets/{dataset}/samples"
    webURL = request.urlopen(url)
    data = webURL.read()
    encoding = webURL.info().get_content_charset('utf-8')
    contents = json.loads(data.decode(encoding))
    if not os.path.exists(os.path.join(path,dataset)):
        os.makedirs(os.path.join(path,dataset))
    with open(os.path.join(path,dataset,'sample_metadata.json'), 'w') as f:
        json.dump(contents, f)
    return contents

datasets = df_subset["experiment.ShortName"].unique()
error_ds = []

for dataset in tqdm(datasets):
    try:
        get_sample_metadata(dataset, path)
    except:
        error_ds.append(dataset)

length =  len(df_subset)
df_subset = df_subset.loc[~df_subset["experiment.ShortName"].isin(error_ds)]
print(length - len(df_subset), "removed rows due to missing sample data")

### save

df_subset.to_csv(os.path.join(data_dir, "metadata", "gemma_dump_FINAL.tsv"), sep='\t', index=False)

### Preliminary stats

#### functions

def flatten_list(l):
    new_l = []
    for item in l:
        if ", " not in item:
            new_l.append(item)
        else:
            for subitem in item.split(", "):
                new_l.append(subitem)
    return new_l

def make_stats(category, overall_df):
    df = overall_df.loc[overall_df["class.Name"] == category]
    term_list = list(df['term.Name'].unique())
    datasets = {
        term : list(df.loc[df['term.Name'] == term]["experiment.ShortName"].unique())
        for term in term_list
        if len(df.loc[df['term.Name'] == term]["experiment.ShortName"].unique()) >= 5
    }  
    platforms = {
        term : flatten_list(list(df.loc[df['term.Name'] == term]["platform.ShortName"].unique()))
        for term in term_list
    }
    term_dict = {
        term : {
            'number_datasets' : len(ds), 
            'number_platforms': len(platforms[term])
        }
        for term, ds in datasets.items()
    }
    term_dict = dict(sorted(term_dict.items(), key=lambda item: item[1]['number_datasets'], reverse = True))
    
    stats_df = pd.DataFrame.from_dict(term_dict, orient = "index")
    stats_df.index.name = category
    return stats_df

#### tissue stats

tissue_stats_from_all = make_stats("organism part", df_subset)
tissue_stats_from_all.head()

tissue_stats_from_human = make_stats("organism part", df_human_only)
tissue_stats_from_human.head()

tissue_stats_from_topplatform = make_stats("organism part", df_GPL570)
tissue_stats_from_topplatform.head()

#### cell line stats

cell_line_stats_from_all = make_stats("cell line", df_subset)
cell_line_stats_from_all.head()

cell_line_stats_from_human = make_stats("cell line", df_human_only)
cell_line_stats_from_human.head()

cell_line_stats_from_topplatform = make_stats("cell line", df_GPL570)
cell_line_stats_from_topplatform.head()

#### cell type stats

cell_type_stats_from_all = make_stats("cell type", df_subset)
cell_type_stats_from_all.head()

cell_type_stats_from_human = make_stats("cell type", df_human_only)
cell_type_stats_from_human.head()

cell_type_stats_from_topplatform = make_stats("cell type", df_GPL570)
cell_type_stats_from_topplatform.head()
