==> analyses/0.2-arrange_metadata_for_curation.ipynb <==
# Arrange metadata for ContNeXt

import getpass
import json
import os
import re
import sys
import time
from difflib import get_close_matches

import pandas as pd
from tqdm import tqdm

getpass.getuser()

sys.version

time.asctime()

# replace here the location of the external data dir
data_dir = os.path.join(os.path.expanduser("~"), "contnext_data", "data")

### open processed Gemma data

df = pd.read_table(os.path.join(data_dir, "metadata", "gemma_dump_FINAL.tsv"))
df

ontology_url_dict = {
    "organism part" : {None : pd.NA},
    "cell type" : {None : pd.NA},
    "cell line" : {None : pd.NA},
}

for i, row in df.drop_duplicates(['class.Name','term.URL','term.Name'])[['class.Name','term.URL','term.Name']].iterrows():
    if row['term.Name'] not in ontology_url_dict[row['class.Name']]:
        ontology_url_dict[row['class.Name']][row['term.Name']] = row['term.URL']

df_human_only = df.loc[df["taxon.Name"] == "human"]
df_GPL570 = df[df['platform.ShortName'].str.contains("GPL570")]

### Process data, arrange metadata, prep manual curation check

#### functions

def is_control(sample_data):
    """ TODO
    this function is adapted from R code given by the PavlidisLab team behind Gemma"""
    
    
    # The recommended version strictly limits to terms that are present in ontologies Gemma supports.
    # The commented out section is for IFF you want to include free text terms that are not in ontology form.
    # Note: Use at own risk.
    controlPatternQuery = '|'.join(['control',
                                    'reference substance role',
                                    'reference subject role',
                                    'baseline participant role',
                                    'wild type genotype',
                                    'initial time point',
    #                                'phosphate buffered saline', #see note
    #                                'control group',  #see note
    #                                'normal', #see note
    #                                'untreated', #see note
    #                                'baseline', #see note
    #                                'wild type', #see note
                                   ])



    positiveFlag = False
    
    factorValues = sample_data["sample"]["factorValues"].values()
    samplePositiveCount = [None]*len(factorValues)
    for i, fvObject in enumerate(factorValues):
        #print(fvObject)
        
        if not fvObject:
                continue
                
        valueGrep = re.search(controlPatternQuery, fvObject) is not None
        
        if valueGrep:
            samplePositiveCount[i] = True

    if any(samplePositiveCount):
        return True
    return False


def get_sample_acc(sample_data):
    return sample_data["accession"]["accession"]


def get_alt_sample_acc(sample_data):
    return sample_data["sample"]["name"]
    
    
def get_term_FV(sample_data, term_type):
    match = ""
    FV = sample_data["sample"]["factorValues"]
    for fvObject in FV.values():
        if not fvObject: continue
        fvObject = fvObject.split("::::")
        for part in fvObject:
            if "Category = "+term_type in part:
                if match: 
                    match += ";"+part.split("Value = ")[1]
                else:
                    match = part.split("Value = ")[1]
    if not match or match == "Male/Female; Normal/Diseased":
        match = sample_data["sample"]["characteristicValues"][term_type] #fallback value
    return match


terms_per_ds = {}
for ds in df["experiment.ShortName"].unique():
    #if ds in error_ds: continue 
    subdf = df.loc[df["experiment.ShortName"] == ds]
    terms_per_ds[ds] = {}
    for term in subdf["class.Name"].unique():
        terms_per_ds[ds][term] = list(subdf.loc[subdf["class.Name"] == term]["term.Name"].unique())

for ds in df["experiment.ShortName"].unique():
    rows =  df.loc[df["experiment.ShortName"]== ds]
    if len(set(rows["taxon.Name"])) != 1:
        raise Warning("too many taxa")

path = os.path.join(data_dir, "metadata", "dataset_sample_metadata")

sample_dict = {}
for ds, categories in terms_per_ds.items():
    ETflag = False
    rows =  df.loc[df["experiment.ShortName"]== ds]
    if "ExperimentTag" in list(rows["class.Type"]):
        ETflag = True
    taxon = list(rows["taxon.Name"])[0]
    pltfm = list(rows["platform.ShortName"])[0]
    if ds == "McLean Hippocampus":
        with open(os.path.join(path, '670/sample_metadata.json'), 'r') as f:
            ds_data = json.load(f)
    else:
        with open(os.path.join(path, ds,'sample_metadata.json'), 'r') as f:
            ds_data = json.load(f)
    
    for j, sample in enumerate(ds_data['data']):
        if not is_control(sample):
            continue
        try:
            acc = get_sample_acc(sample)
        except TypeError:
            acc = get_alt_sample_acc(sample)
        sample_dict[acc] = {"dataset" : ds, 
                            "platform" : pltfm,
                            "species" : taxon, 
                            "original organism part" : pd.NA,
                            "organism part" : pd.NA,
                            "organism part URL": pd.NA,
                            "original cell type" : pd.NA,
                            "cell type" : pd.NA, 
                            "cell type URL" : pd.NA,
                            "original cell line" : pd.NA,
                            "cell line" : pd.NA,
                            "cell line URL" : pd.NA,
                            "manual check": "",
                           }  
        if ETflag:
            for term_type, terms in categories.items():
                for term in terms:    
                    sample_dict[acc][term_type] = term
                    sample_dict[acc][term_type+" URL"] = ontology_url_dict[term_type][term]
        else:
            for term_type, terms in categories.items():
                try:
                    term = get_term_FV(sample, term_type)
                except KeyError:
                    continue
                if term == "Male/Female; Normal/Diseased":
                    continue
                sample_dict[acc]["original "+term_type] = term
                match = []
                iters = 0
                similarity_cutoff = .8
                match = get_close_matches(term, terms, 1, similarity_cutoff)
                if not match: #no match found
                    for word in term.split(): #try and break up the words
                        match = get_close_matches(word, terms, 1, similarity_cutoff)
                        if match: 
                            sample_dict[acc]["manual check"] += f"{j} word split on {term_type};"
                            break
                if not match: #still no match
                    for word in term.split(";"): #try and break up parts
                        match = get_close_matches(word, terms, 1, similarity_cutoff)
                        if match: 
                            sample_dict[acc]["manual check"] += f"{j} semicolon split on {term_type};"
                            break
                while len(match) == 0: #still no match, lower standards for match
                    similarity_cutoff -= .05
                    match = get_close_matches(term, terms, 1, similarity_cutoff)
                    if match and similarity_cutoff >.35:
                        sample_dict[acc]["manual check"] += f"{j} {round(similarity_cutoff, 2)} on {term_type} {terms};"
                    if similarity_cutoff < .35:
                        match = [None]
                        sample_dict[acc]["manual check"] += f"{j} check {term_type} {terms};"
                        break
                sample_dict[acc][term_type] = match[0]
                sample_dict[acc][term_type+" URL"] = ontology_url_dict[term_type][match[0]]
        

sample_df = pd.DataFrame.from_dict(sample_dict, orient="index")
print(len(sample_df))
sample_df

print(len(sample_df["organism part"].unique()), "tissues, totaling in", len(sample_df.loc[sample_df["organism part"].notnull()]), "samples from",  len(sample_df.loc[sample_df["organism part"].notnull()]["dataset"].unique()),"datasets")
print(len(sample_df["cell type"].unique()), "cell type, totaling in", len(sample_df.loc[sample_df["cell type"].notnull()]), "samples from",  len(sample_df.loc[sample_df["cell type"].notnull()]["dataset"].unique()),"datasets")
print(len(sample_df["cell line"].unique()), "cell line, totaling in", len(sample_df.loc[sample_df["cell line"].notnull()]), "samples from",  len(sample_df.loc[sample_df["cell line"].notnull()]["dataset"].unique()),"datasets")

sample_df.to_csv(os.path.join(data_dir, "metadata", "metadata_before_curation.tsv"), sep='\t', index=True)
dropped = sample_df.drop_duplicates()
dropped.to_csv(os.path.join(data_dir, "metadata", "metadata_for_manual_curation.tsv"), sep='\t', index=True)

print(len(dropped))
len(dropped.loc[~dropped["manual check"].isin([""])])

platform_to_keep = "GPL570"

subset_sample_df = sample_df[sample_df['platform'].str.contains(platform_to_keep)]
subset_sample_df.to_csv(os.path.join(data_dir, "metadata", "metadata_before_curation_shortened.tsv"), sep='\t', index=True)
dropped_subset = subset_sample_df.drop_duplicates()
dropped_subset.to_csv(os.path.join(data_dir, "metadata", "metadata_for_manual_curation_shortened.tsv"), sep='\t', index=True)
print(len(dropped_subset))
len(dropped_subset.loc[~dropped_subset["manual check"].isin([""])])
