==> analyses/0.3-analyze_metadata_overview.ipynb <==
# Analyze metadata for ContNeXt

import getpass
import json
import os
import sys
import time

import pandas as pd

getpass.getuser()

sys.version

time.asctime()

# replace here the location of the external data dir
data_dir = os.path.join(os.path.expanduser("~"), "contnext_data", "data")

### metadata for top human platform  + ontology term name mappings

metadata = pd.read_table(os.path.join(data_dir, "metadata", "final_metadata.tsv"), index_col=0)

human_data = metadata[metadata['platform'].str.contains("GPL570")]
human_data = human_data.loc[human_data["species"] == "human"]
human_data.index.name= "sample_id"

with open(os.path.join(data_dir, "mappings", "uberon_name_mappings.json"), 'r') as f:
    uberon_name_mappings = json.load(f)

with open(os.path.join(data_dir, "mappings", "CL_name_mappings.json"), 'r') as f:
    CL_name_mappings = json.load(f)

with open(os.path.join(data_dir, "mappings", "CLO_name_mappings.json"), 'r') as f:
    CLO_name_mappings = json.load(f)

### Term overview

#### tissues

human_tissue_subset = human_data[human_data["organism part"].notnull()]
human_tissue_subset = human_tissue_subset[human_tissue_subset["organism part"] != ""]
human_tissue_subset = human_tissue_subset[human_tissue_subset['organism part URL'].str.contains("UBERON")]

human_tissue_subset

# make a dict to keep track of number of samples per uberon_id

samples_per_uberon = {(k.split("_")[-1]).strip() : v for k,v in dict(human_tissue_subset["organism part URL"].value_counts()).items()}

# make a dict to keep track of number of datasets/experiments per uberon_id

exp_per_uberon = {(k.split("_")[-1]).strip() : v for k,v in dict(human_tissue_subset.groupby('organism part URL').apply(lambda x: len(x['dataset'].unique()))).items()}

with open(os.path.join(data_dir, "misc_data", "FULL_tissue_overview.tsv"),"w") as f:
    f.write("UBERON_id\ttissue_name\tnumber_experiments\tnumber_samples\n")
    for uberon_id, _ in sorted(samples_per_uberon.items(), reverse=True, key=lambda item: item[1]):
        f.write(f"UBERON:{uberon_id}\t{uberon_name_mappings[uberon_id]}\t{exp_per_uberon[uberon_id]}\t{samples_per_uberon[uberon_id]}\n")

#### cell types

human_celltype_subset = human_data[human_data["cell type"].notnull()]
human_celltype_subset = human_celltype_subset[human_celltype_subset["cell type"] != ""]
human_celltype_subset = human_celltype_subset[human_celltype_subset['cell type URL'].str.contains("CL_")]

# make a dict to keep track of number of samples per CL_id

samples_per_CL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_celltype_subset["cell type URL"].value_counts()).items()}

# make a dict to keep track of number of datasets/experiments per CL_id

exp_per_CL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_celltype_subset.groupby('cell type URL').apply(lambda x: len(x['dataset'].unique()))).items()}

with open(os.path.join(data_dir, "misc_data", "FULL_celltype_overview.tsv"),"w") as f:
    f.write("CL_id\tcell_type_name\tnumber_experiments\tnumber_samples\n")
    for CL_id, _ in sorted(exp_per_CL.items(), reverse=True, key=lambda item: item[1]):
        f.write(f"CL:{CL_id}\t{CL_name_mappings[CL_id]}\t{exp_per_CL[CL_id]}\t{samples_per_CL[CL_id]}\n")

#### cell lines

human_cellline_subset = human_data[human_data["cell line"].notnull()]
human_cellline_subset = human_cellline_subset[human_cellline_subset["cell line"] != ""]
human_cellline_subset = human_cellline_subset[human_cellline_subset['cell line URL'].str.contains("CLO_")]

# make a dict to keep track of number of samples per CLO id

samples_per_CLO = {(k.split("_")[-1]).strip() : v for k,v in dict(human_cellline_subset["cell line URL"].value_counts()).items()}

# make a dict to keep track of number of datasets/experiments per CLO_id

exp_per_CLO = {(k.split("_")[-1]).strip() : v for k,v in dict(human_cellline_subset.groupby('cell line URL').apply(lambda x: len(x['dataset'].unique()))).items()}

CLO_name_mappings = {}
for i, row in human_cellline_subset.iterrows():
    clo_id = row["cell line URL"].split("_")[-1]
    if clo_id not in CLO_name_mappings:
        CLO_name_mappings[clo_id] = row["cell line"]

with open(os.path.join(data_dir, "misc_data", "FULL_cellline_overview.tsv"),"w") as f:
    f.write("CLO_id\tcell_line_name\tnumber_experiments\tnumber_samples\n")
    for CLO_id, _ in sorted(exp_per_CLO.items(), reverse=True, key=lambda item: item[1]):
        f.write(f"CLO:{CLO_id}\t{CLO_name_mappings[CLO_id]}\t{exp_per_CLO[CLO_id]}\t{samples_per_CLO[CLO_id]}\n")
