==> analyses/0.5-prep_metadata_for_net_construction.ipynb <==
# Prepare metadata for the R scripts to build the co-expression networks

import getpass
import json
import os
import sys
import time

import pandas as pd

getpass.getuser()

sys.version

time.asctime()

# replace here the location of the external data dir
data_dir = os.path.join(os.path.expanduser("~"), "contnext_data", "data")

___

tissues = pd.read_table(os.path.join(data_dir, "misc_data", "FULL_tissue_overview.tsv"))

tissues_afterDL = pd.read_table(os.path.join(data_dir, "misc_data", "FULL_tissue_overview_after_download.tsv"))

celltypes = pd.read_table(os.path.join(data_dir, "misc_data", "FULL_celltype_overview.tsv"))

celltypes_afterDL = pd.read_table(os.path.join(data_dir, "misc_data", "FULL_celltype_overview_after_download.tsv"))

celllines = pd.read_table(os.path.join(data_dir, "misc_data", "FULL_cellline_overview.tsv"))

celllines_afterDL = pd.read_table(os.path.join(data_dir, "misc_data", "FULL_cellline_overview_after_download.tsv"))

___

tissues

tis_terms = tissues[(tissues["number_experiments"] >= 3) & (tissues["number_samples"] >= 20)]["tissue_name"]
tis_terms_afterDL = tissues_afterDL[(tissues_afterDL["number_experiments"] >= 3) & (tissues_afterDL["number_samples"] >= 20)]["tissue_name"]

num_tissues = len(tis_terms)
num_tissues_afterDL = len(tis_terms_afterDL)

print(f"tissues meeting criteria in top platform: {num_tissues}, after download(top platform): {num_tissues_afterDL}")

celltype_terms = celltypes[(celltypes["number_experiments"] >= 3) & (celltypes["number_samples"] >= 20)]["cell_type_name"]
celltype_terms_afterDL = celltypes_afterDL[(celltypes_afterDL["number_experiments"] >= 3) & (celltypes_afterDL["number_samples"] >= 20)]["cell_type_name"]

num_celltypes = len(celltype_terms)
num_celltypes_afterDL = len(celltype_terms_afterDL)

print(f"cell types meeting criteria in top platform: {num_celltypes}, after download(top platform): {num_celltypes_afterDL}")

cellline_terms = celllines[(celllines["number_experiments"] >= 3) & (celllines["number_samples"] >= 20)]["cell_line_name"]
cellline_terms_afterDL = celllines_afterDL[(celllines_afterDL["number_experiments"] >= 3) & (celllines_afterDL["number_samples"] >= 20)]["cell_line_name"]

num_celllines = len(cellline_terms)
num_celllines_afterDL = len(cellline_terms_afterDL)

print(f"cell lines meeting criteria in top platform: {num_celllines}, after download(top platform): {num_celllines_afterDL}")

metadata = pd.read_table(os.path.join(data_dir, "metadata", "final_metadata.tsv"), index_col=0)
metadata.index.name= "sample_id"

metadata_afterDL = pd.read_table(os.path.join(data_dir, "metadata", "metadataFinal_afterDataLoading.tsv"), index_col=0)
metadata_afterDL.index.name= "sample_id"

#human_data = metadata[metadata['platform'].str.contains("GPL570")]
#human_data = human_data.loc[human_data["species"] == "human"]

human_data_after_DL = metadata_afterDL[metadata_afterDL['platform'].str.contains("GPL570")]
human_data_after_DL = human_data_after_DL.loc[human_data_after_DL["species"] == "human"]

def prep_metadata_for_nets(samples_per_id, exp_per_id, metaData, term_type, name_mappings):

    path = os.path.join(data_dir, "data_for_coexp_network_construction", term_type)

    group_name_list = []
    group_id_list = []
    total_sample_num = 0
    total_dataset_num = 0
    datasets_total = set()
    for ID, num in samples_per_id.items():
        if num < 20 : continue # threshold
        if exp_per_id[ID] < 3 : continue # threshold

        os.makedirs(os.path.join(path,ID), exist_ok=True)

        #save group-specific metadata
        metaD = metaData.loc[metaData[term_type+".URL"].str.contains(ID)]
        metaD.to_csv(os.path.join(path, ID, "metadata.tsv"), sep="\t", index=True)

        #save list of datasets that contain specific group
        datasets = list(metaD["dataset"].unique())

        with open(os.path.join(path,ID,"datasets.txt"), "w") as f:
            f.write("\n".join(datasets))
            f.write("\n")

        total_sample_num += num
        for ds in datasets:
            if ds not in datasets_total:
                total_dataset_num += 1
                datasets_total.add(ds)

        group_name_list.append(name_mappings[ID])
        group_id_list.append(ID)
    print(f"{total_sample_num} {term_type} samples from {total_dataset_num} datasets with {len(group_name_list)} groups")
    return group_name_list, group_id_list, datasets_total


### tissue

with open(os.path.join(data_dir, "mappings", "uberon_name_mappings.json"), 'r') as f:
    uberon_name_mappings = json.load(f)

#human_tissue_subset = human_data[human_data["organism part"].notnull()]
#human_tissue_subset = human_tissue_subset[human_tissue_subset["organism part"] != ""]
#human_tissue_subset = human_tissue_subset[human_tissue_subset['organism part URL'].str.contains("UBERON")]


human_tissue_subset_afterDL = human_data_after_DL[human_data_after_DL["organism.part"].notnull()]
human_tissue_subset_afterDL = human_tissue_subset_afterDL[human_tissue_subset_afterDL["organism.part"] != ""]
human_tissue_subset_afterDL = human_tissue_subset_afterDL[human_tissue_subset_afterDL['organism.part.URL'].str.contains("UBERON")]

# make a dict to keep track of number of samples per uberon_id

#samples_per_uberon = {(k.split("_")[-1]).strip() : v for k,v in dict(human_tissue_subset["organism part URL"].value_counts()).items()}
samples_per_uberon_afterDL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_tissue_subset_afterDL["organism.part.URL"].value_counts()).items()}

# make a dict to keep track of number of datasets/experiments per uberon_id

#exp_per_uberon = {(k.split("_")[-1]).strip() : v for k,v in dict(human_tissue_subset.groupby('organism part URL').apply(lambda x: len(x['dataset'].unique()))).items()}
exp_per_uberon_afterDL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_tissue_subset_afterDL.groupby('organism.part.URL').apply(lambda x: len(x['dataset'].unique()))).items()}

tissue_group_list, tissue_id_list, tissue_ds_list = prep_metadata_for_nets(samples_per_uberon_afterDL, exp_per_uberon_afterDL, human_tissue_subset_afterDL, "organism.part", uberon_name_mappings)

tissue_id_list

## cell type

with open(os.path.join(data_dir, "mappings", "CL_name_mappings.json"), 'r') as f:
    CL_name_mappings = json.load(f)

#human_celltype_subset = human_data[human_data["cell type"].notnull()]
#human_celltype_subset = human_celltype_subset[human_celltype_subset["cell type"] != ""]
#human_celltype_subset = human_celltype_subset[human_celltype_subset['cell type URL'].str.contains("CL_")]

human_celltype_subset_after_DL = human_data_after_DL[human_data_after_DL["cell.type"].notnull()]
human_celltype_subset_after_DL = human_celltype_subset_after_DL[human_celltype_subset_after_DL["cell.type"] != ""]
human_celltype_subset_after_DL = human_celltype_subset_after_DL[human_celltype_subset_after_DL['cell.type.URL'].str.contains("CL_")]

# make a dict to keep track of number of samples per CL_id

#samples_per_CL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_celltype_subset["cell type URL"].value_counts()).items()}
samples_per_CL_afterDL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_celltype_subset_after_DL["cell.type.URL"].value_counts()).items()}

# make a dict to keep track of number of datasets/experiments per CL_id

#exp_per_CL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_celltype_subset.groupby('cell type URL').apply(lambda x: len(x['dataset'].unique()))).items()}
exp_per_CL_afterDL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_celltype_subset_after_DL.groupby('cell.type.URL').apply(lambda x: len(x['dataset'].unique()))).items()}


ct_group_list, ct_id_list, ct_ds_list = prep_metadata_for_nets(samples_per_CL_afterDL, exp_per_CL_afterDL, human_celltype_subset_after_DL, "cell.type", CL_name_mappings)

## cell line 

#human_cellline_subset = human_data[human_data["cell line"].notnull()]
#human_cellline_subset = human_cellline_subset[human_cellline_subset["cell line"] != ""]
#human_cellline_subset = human_cellline_subset[human_cellline_subset['cell line URL'].str.contains("CLO_")]

human_cellline_subset_afterDL = human_data_after_DL[human_data_after_DL["cell.line"].notnull()]
human_cellline_subset_afterDL = human_cellline_subset_afterDL[human_cellline_subset_afterDL["cell.line"] != ""]
human_cellline_subset_afterDL = human_cellline_subset_afterDL[human_cellline_subset_afterDL['cell.line.URL'].str.contains("CLO_")]

# make a dict to keep track of number of samples per CLO id

#samples_per_CLO = {(k.split("_")[-1]).strip() : v for k,v in dict(human_cellline_subset["cell line URL"].value_counts()).items()}
samples_per_CLO_afterDL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_cellline_subset_afterDL["cell.line.URL"].value_counts()).items()}

# make a dict to keep track of number of datasets/experiments per CLO_id

#exp_per_CLO = {(k.split("_")[-1]).strip() : v for k,v in dict(human_cellline_subset.groupby('cell line URL').apply(lambda x: len(x['dataset'].unique()))).items()}
exp_per_CLO__afterDL = {(k.split("_")[-1]).strip() : v for k,v in dict(human_cellline_subset_afterDL.groupby('cell.line.URL').apply(lambda x: len(x['dataset'].unique()))).items()}

CLO_name_mappings = {}
for i, row in human_cellline_subset_afterDL.iterrows():
    clo_id = row["cell.line.URL"].split("_")[-1]
    if clo_id not in CLO_name_mappings:
        CLO_name_mappings[clo_id] = row["cell.line"]

cl_group_list, cl_id_list, cl_ds_list = prep_metadata_for_nets(samples_per_CLO_afterDL, exp_per_CLO__afterDL, human_cellline_subset_afterDL, "cell.line", CLO_name_mappings)
